+++ core/src/test/java/arez/AbstractArezTest.java
@@ -26,6 +26,7 @@ public abstract class AbstractArezTest
     BrainCheckTestUtil.resetConfig( false );
     ArezTestUtil.resetConfig( false );
     ArezTestUtil.enableZones();
+    ArezTestUtil.resetState();
     getProxyLogger().setLogger( new TestLogger() );
     _ignoreObserverErrors = false;
     _printObserverErrors = true;