+++ integration-tests/src/test/java/arez/integration/MultiZoneIntegrationTest.java
@@ -20,6 +20,7 @@ public class MultiZoneIntegrationTest
     throws Throwable
   {
     ArezTestUtil.enableZones();
+    ArezTestUtil.resetState();
 
     final Zone zone1 = Arez.createZone();
     final Zone zone2 = Arez.createZone();
@@ -61,6 +62,7 @@ public class MultiZoneIntegrationTest
     throws Throwable
   {
     ArezTestUtil.enableZones();
+    ArezTestUtil.resetState();
 
     final Zone zone1 = Arez.createZone();
     final Zone zone2 = Arez.createZone();